+++ API/src/main/java/fr/maxlego08/essentials/api/dto/UserDTO.java
@@ -10,5 +10,5 @@ import java.util.UUID;
  */
 public record UserDTO(UUID unique_id, String name, String last_location, Integer ban_sanction_id,
                       Integer mute_sanction_id, long play_time, Date created_at, Date updated_at, long vote, long vote_offline
-                        , Boolean frozen) {
+                        , Boolean frozen, long fly_seconds) {
 }